api/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
psycopg2-binary==2.9.10
